Agrega sistema de creación de torres

+++ Assets/_project/Code/Core/EnemyDataSO.cs
@@ -13,7 +13,7 @@ public class EnemyDataSO : ScriptableObject
 
     [Header("Rewards")]
     public int goldReward;
-    public int damageToCore;
+    public float damageToCore;
 
     [Header("Combat")]
     public float attackResistance;
